refactor: type 및 use-case 사용하는 모듈 하위로 이동
