+++ src/services/counselings/applications/counsel-managements/counsel-managements.module.ts
@@ -1,5 +1,5 @@
 import { CounselManagementsFacade } from "~counselings/applications/counsel-managements/counsel-managements.facade";
-import { ProceedCounselingUseCase } from "~counselings/applications/use-cases/proceed-counseling";
+import { ProceedCounselingUseCase } from "~counselings/applications/counsel-managements/use-cases/proceed-counseling";
 import { CounselMessagesModule } from "~counselings/domains/counselMessages/counselMessages.module";
 import { CounselorsModule } from "~counselings/domains/counselors/counselors.module";
 import { CounselsModule } from "~counselings/domains/counsels/counsels.module";
