+++ src/services/counselings/applications/counsel-prompt-managements/use-cases/dtos/validate-prompt-version.dto.ts
